refactor: 동아리 폼 화면 로직 분리
